requirements.txt
streamlit
pandas
pillow
numpy